Séparation en différentes classes Fix #11
